remove thread factory from services and move it to transport and fix name format in one place and use the factory in relevant places
