+++ cluster/src/main/java/io/scalecube/cluster/fdetector/FailureDetector.java
@@ -8,6 +8,7 @@ import io.scalecube.cluster.membership.MemberStatus;
 import io.scalecube.cluster.membership.MembershipEvent;
 import io.scalecube.transport.Transport;
 import io.scalecube.transport.Message;
+import io.scalecube.transport.ThreadFactory;
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
 
@@ -88,8 +89,7 @@ public final class FailureDetector implements IFailureDetector {
     this.membership = membership;
     this.config = config;
     String nameFormat = "sc-fdetector-" + transport.address().toString();
-    this.executor = Executors.newSingleThreadScheduledExecutor(
-        new ThreadFactoryBuilder().setNameFormat(nameFormat).setDaemon(true).build());
+    this.executor = ThreadFactory.singleScheduledExecutorService(nameFormat);
     this.scheduler = Schedulers.from(executor);
   }
